requirements.txt
discord.py
flask
requests
asyncio
quart